Move EE broker startup to startup folder
