+++ ee/server/local-services/ldap/service.ts
@@ -1,5 +1,3 @@
-import '../../broker';
-
 import { LDAPEEManager } from '../../lib/ldap/Manager';
 import { ILDAPEEService } from '../../sdk/types/ILDAPEEService';
 import { ServiceClassInternal } from '../../../../server/sdk/types/ServiceClass';
